ambilight fixed

+++ src/bridge.cpp
@@ -580,10 +580,10 @@ void Bridge::setAmbientLight(int blue, int green, int red, int white, int durati
     LOG(moba::INFO) << "setAmbientLight <" << blue << "><" << green << "><" << red << "><" << white << "><" << duration << ">" << std::endl;
     std::stringstream ss;
     ss <<
-        4096.0 * blue / 100 << ";" <<
-        4096.0 * green / 100 << ";" <<
-        4096.0 * red / 100 << ";" <<
-        4096.0 * white / 100 << ";" << duration;
+        static_cast<int>(4096.0 * blue / 100) << ";" <<
+        static_cast<int>(4096.0 * green / 100) << ";" <<
+        static_cast<int>(4096.0 * red / 100) << ";" <<
+        static_cast<int>(4096.0 * white / 100) << ";" << duration;
     LOG(moba::INFO) << "sending " << ss.str() << std::endl;
     ipc->send(ss.str(), moba::IPC::CMD_RUN);
 }
